requirements.txt
# AWS CDK and Core Dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.100.0a0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# CLI and Configuration
click>=8.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typer>=0.9.0
rich>=13.0.0

# Web Framework for Portal
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Security and Authentication
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Monitoring and Observability
prometheus-client>=0.17.0
structlog>=23.1.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-exporter-otlp>=1.19.0

# Data Processing and ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Utilities
requests>=2.31.0
httpx>=0.24.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
toml>=0.10.2
jsonschema>=4.18.0

# Template Engine
cookiecutter>=2.2.0
jinja2-time>=0.2.0

# Git Integration
GitPython>=3.1.32

# Testing (included in main requirements for CI/CD)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
moto[all]>=4.2.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pylint>=2.17.0

# Security Scanning
bandit>=1.7.5
safety>=2.3.0
semgrep>=1.35.0
checkov>=2.3.0
detect-secrets>=1.4.0
pip-audit>=2.6.0
pip-licenses>=4.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0
mkdocs-mermaid2-plugin>=1.1.0
mkdocstrings>=0.22.0
mkdocstrings-python>=1.5.0
